CardsParser reworked

+++ src/com/github/tyrantsim/jtuo/parsers/SkillsSetParser.java
@@ -433,7 +433,7 @@ public class SkillsSetParser {
     private static void setSkillSpec(Card card, String x, String y, String all, String c, String card_id, String trigger, SkillSpec new_skill, String n, String s, String s2) {
         new_skill.setAll(all != null && all.equals("1"));
         if (!x.isEmpty()) {
-            new_skill.setX(Float.parseFloat(x));
+            new_skill.setX(Integer.parseInt(x));
         }
         if (!y.isEmpty()) {
             new_skill.setY(Faction.values()[Integer.parseInt(y)]);
